+++ CnCTDRAMapEditor/Model/InfantryType.cs
@@ -30,28 +30,29 @@ namespace MobiusEditor.Model
         public string DisplayName { get; private set; }
         public string OwnerHouse { get; private set; }
         public UnitTypeFlag Flag { get; private set; }
-        public bool IsArmed => this.Flag.HasFlag(UnitTypeFlag.IsArmed);
+        public bool IsArmed => Flag.HasFlag(UnitTypeFlag.IsArmed);
         public bool IsAircraft => false;
         public bool IsFixedWing => false;
-        public bool IsExpansionOnly => this.Flag.HasFlag(UnitTypeFlag.IsExpansionUnit);
+        public bool IsExpansionOnly => Flag.HasFlag(UnitTypeFlag.IsExpansionUnit);
         public bool IsHarvester => false;
-        public bool CanRemap => !this.Flag.HasFlag(UnitTypeFlag.NoRemap);
+        public bool CanRemap => !Flag.HasFlag(UnitTypeFlag.NoRemap);
         public string ClassicGraphicsSource { get; set; }
-        public Byte[] ClassicGraphicsRemap { get; set; }
+        public byte[] ClassicGraphicsRemap { get; set; }
+        public bool GraphicsFound { get; private set; }
 
         public Bitmap Thumbnail { get; set; }
         private string nameId;
 
         public InfantryType(int id, string name, string textId, string ownerHouse, string remappedFrom, byte[] remapTable, UnitTypeFlag flags)
         {
-            this.ID = id;
-            this.Name = name;
-            this.GraphicsSource = name;
-            this.nameId = textId;
-            this.OwnerHouse = ownerHouse;
-            this.Flag = flags;
-            this.ClassicGraphicsSource = remappedFrom;
-            this.ClassicGraphicsRemap = remapTable;
+            ID = id;
+            Name = name;
+            GraphicsSource = name;
+            nameId = textId;
+            OwnerHouse = ownerHouse;
+            Flag = flags;
+            ClassicGraphicsSource = remappedFrom;
+            ClassicGraphicsRemap = remapTable;
         }
 
         public InfantryType(int id, string name, string textId, string ownerHouse, UnitTypeFlag flags)
@@ -71,19 +72,19 @@ namespace MobiusEditor.Model
             }
             else if (obj is sbyte sb)
             {
-                return this.ID == sb;
+                return ID == sb;
             }
             else if (obj is byte b)
             {
-                return this.ID == b;
+                return ID == b;
             }
             else if (obj is int i)
             {
-                return this.ID == i;
+                return ID == i;
             }
             else if (obj is string)
             {
-                return string.Equals(this.Name, obj as string, StringComparison.OrdinalIgnoreCase);
+                return string.Equals(Name, obj as string, StringComparison.OrdinalIgnoreCase);
             }
 
             return base.Equals(obj);
@@ -91,35 +92,35 @@ namespace MobiusEditor.Model
 
         public override int GetHashCode()
         {
-            return this.ID.GetHashCode();
+            return ID.GetHashCode();
         }
 
         public override string ToString()
         {
-            return this.Name;
+            return Name;
         }
 
 
         public void InitDisplayName()
         {
-            this.DisplayName = !String.IsNullOrEmpty(this.nameId) && !String.IsNullOrEmpty(Globals.TheGameTextManager[this.nameId])
-                ? Globals.TheGameTextManager[this.nameId] + " (" + this.Name.ToUpperInvariant() + ")"
-                : this.Name.ToUpperInvariant();
+            DisplayName = !String.IsNullOrEmpty(nameId) && !String.IsNullOrEmpty(Globals.TheGameTextManager[nameId])
+                ? Globals.TheGameTextManager[nameId] + " (" + Name.ToUpperInvariant() + ")"
+                : Name.ToUpperInvariant();
         }
 
         public void Init(HouseType house, DirectionType direction)
         {
-            this.InitDisplayName();
-            Bitmap oldImage = this.Thumbnail;
+            InitDisplayName();
+            Bitmap oldImage = Thumbnail;
             // Initialisation for the special RA civilian remapping logic. Normally, code should never check the type of the tileset manager
             // and use a specific one, but this is a special case that'd be more convoluted to get around by implementing it generally.
-            if (((this.ClassicGraphicsSource != null && !String.Equals(this.Name, this.ClassicGraphicsSource, StringComparison.OrdinalIgnoreCase))
-                || this.ClassicGraphicsRemap != null) && Globals.TheTilesetManager is TilesetManagerClassic tsmc)
+            if (((ClassicGraphicsSource != null && !String.Equals(Name, ClassicGraphicsSource, StringComparison.OrdinalIgnoreCase))
+                || ClassicGraphicsRemap != null) && Globals.TheTilesetManager is TilesetManagerClassic tsmc)
             {
-                string actualSprite = this.ClassicGraphicsSource ?? this.Name;
+                string actualSprite = ClassicGraphicsSource ?? Name;
                 // Use special override that 100% makes sure previously-cached versions are cleared, so previous accidental fetches do not corrupt it.
-                this.GraphicsSource = actualSprite + " (override for infantry "+ this.Name.ToUpperInvariant() + ")";
-                tsmc.GetTeamColorTileData(this.GraphicsSource, 0, null, out _, true, false, true, actualSprite, this.ClassicGraphicsRemap, true);
+                GraphicsSource = actualSprite + " (override for infantry "+ Name.ToUpperInvariant() + ")";
+                tsmc.GetTeamColorTileData(GraphicsSource, 0, null, out _, true, false, true, actualSprite, ClassicGraphicsRemap, true);
             }
             Infantry mockInfantry = new Infantry(null)
             {
@@ -137,8 +138,9 @@ namespace MobiusEditor.Model
                 {
                     render.RenderAction(g);
                 }
+                GraphicsFound = !render.IsDummy;
             }
-            this.Thumbnail = infantryThumbnail;
+            Thumbnail = infantryThumbnail;
             if (oldImage != null)
             {
                 try { oldImage.Dispose(); }
@@ -147,8 +149,8 @@ namespace MobiusEditor.Model
         }
         public void Reset()
         {
-            Bitmap oldImage = this.Thumbnail;
-            this.Thumbnail = null;
+            Bitmap oldImage = Thumbnail;
+            Thumbnail = null;
             if (oldImage != null)
             {
                 try { oldImage.Dispose(); }