+++ CnCTDRAMapEditor/Model/TerrainType.cs
@@ -27,11 +27,11 @@ namespace MobiusEditor.Model
         public int ID { get; private set; }
         public string Name { get; private set; }
         public string DisplayName { get; private set; }
-        public Rectangle OverlapBounds => new Rectangle(Point.Empty, this.Size);
+        public Rectangle OverlapBounds => new Rectangle(Point.Empty, Size);
         public bool[,][] OverlapMask { get; private set; }
         public bool[,][] ContentMask { get; private set; }
         public bool[,] OccupyMask { get; private set; }
-        public Size Size => this.OccupyMask.GetDimensions();
+        public Size Size => OccupyMask.GetDimensions();
         public bool[,] BaseOccupyMask => OccupyMask;
         public int ZOrder => Globals.ZOrderDefault;
         public bool ExistsInTheater { get; private set; }
@@ -45,6 +45,7 @@ namespace MobiusEditor.Model
         public bool IsExpansionOnly => false;
         public bool CanRemap => false;
         public Point CenterPoint { get; private set; }
+        public bool GraphicsFound { get; private set; }
         private string nameId;
 
         public Bitmap Thumbnail { get; set; }
@@ -64,18 +65,18 @@ namespace MobiusEditor.Model
         /// <param name="placementLand">Land type this should be placed down on. Currently unused.</param>
         public TerrainType(int id, string name, string textId, int width, int height, Point centerPoint, string occupyMask, string graphicsSource, int displayIcon, LandType placementLand)
         {
-            this.ID = id;
-            this.Name = name;
-            this.nameId = textId;
-            this.OccupyMask = GeneralUtils.GetMaskFromString(width, height, occupyMask, '0', ' ');
+            ID = id;
+            Name = name;
+            nameId = textId;
+            OccupyMask = GeneralUtils.GetMaskFromString(width, height, occupyMask, '0', ' ');
             if (centerPoint == Point.Empty)
             {
                 centerPoint = GeneralUtils.GetOccupiedCenter(OccupyMask, new Size(Globals.PixelWidth, Globals.PixelHeight));
             }
-            this.CenterPoint = centerPoint;
-            this.GraphicsSource = graphicsSource == null ? name : graphicsSource;
-            this.DisplayIcon = displayIcon;
-            this.PlacementLand = placementLand;
+            CenterPoint = centerPoint;
+            GraphicsSource = graphicsSource == null ? name : graphicsSource;
+            DisplayIcon = displayIcon;
+            PlacementLand = placementLand;
         }
 
         /// <summary>
@@ -161,11 +162,11 @@ namespace MobiusEditor.Model
             }
             else if (obj is sbyte)
             {
-                return this.ID == (sbyte)obj;
+                return ID == (sbyte)obj;
             }
             else if (obj is string)
             {
-                return string.Equals(this.Name, obj as string, StringComparison.OrdinalIgnoreCase);
+                return string.Equals(Name, obj as string, StringComparison.OrdinalIgnoreCase);
             }
 
             return base.Equals(obj);
@@ -173,30 +174,30 @@ namespace MobiusEditor.Model
 
         public override int GetHashCode()
         {
-            return this.ID.GetHashCode();
+            return ID.GetHashCode();
         }
 
         public override string ToString()
         {
-            return (this.Name ?? String.Empty).ToUpperInvariant();
+            return (Name ?? String.Empty).ToUpperInvariant();
         }
 
         public void InitDisplayName()
         {
-            this.DisplayName = !String.IsNullOrEmpty(this.nameId) && !String.IsNullOrEmpty(Globals.TheGameTextManager[this.nameId])
-                ? Globals.TheGameTextManager[this.nameId] + " (" + this.Name.ToUpperInvariant() + ")"
-                : this.Name.ToUpperInvariant();
+            DisplayName = !String.IsNullOrEmpty(nameId) && !String.IsNullOrEmpty(Globals.TheGameTextManager[nameId])
+                ? Globals.TheGameTextManager[nameId] + " (" + Name.ToUpperInvariant() + ")"
+                : Name.ToUpperInvariant();
         }
 
         public void Init()
         {
-            this.InitDisplayName();
-            this.ExistsInTheater = Globals.TheTilesetManager.TileExists(this.GraphicsSource);
-            if (!this.ExistsInTheater && !String.Equals(this.GraphicsSource, this.Name, StringComparison.InvariantCultureIgnoreCase))
+            InitDisplayName();
+            ExistsInTheater = Globals.TheTilesetManager.TileExists(GraphicsSource);
+            if (!ExistsInTheater && !String.Equals(GraphicsSource, Name, StringComparison.InvariantCultureIgnoreCase))
             {
-                this.ExistsInTheater = Globals.TheTilesetManager.TileExists(this.Name);
+                ExistsInTheater = Globals.TheTilesetManager.TileExists(Name);
             }
-            Bitmap oldImage = this.Thumbnail;
+            Bitmap oldImage = Thumbnail;
             Terrain mockTerrain = new Terrain()
             {
                 Type = this,
@@ -204,22 +205,23 @@ namespace MobiusEditor.Model
             RenderInfo render = MapRenderer.RenderTerrain(Point.Empty, Globals.PreviewTileSize, Globals.PreviewTileScale, mockTerrain, false);
             if (render.RenderedObject != null)
             {
-                Bitmap th = new Bitmap(this.Size.Width * Globals.PreviewTileSize.Width, this.Size.Height * Globals.PreviewTileSize.Height);
+                Bitmap th = new Bitmap(Size.Width * Globals.PreviewTileSize.Width, Size.Height * Globals.PreviewTileSize.Height);
                 th.SetResolution(96, 96);
                 using (Graphics g = Graphics.FromImage(th))
                 {
                     MapRenderer.SetRenderSettings(g, Globals.PreviewSmoothScale);
                     render.RenderAction(g);
                 }
-                this.Thumbnail = th;
+                GraphicsFound = !render.IsDummy;
+                Thumbnail = th;
                 // calculate the areas of this that can overlap other objects (include shadow)
-                this.OverlapMask = GeneralUtils.MakeOpaqueMask(th, this.Size, 25, 10, 20, 0x10, false);
+                OverlapMask = GeneralUtils.MakeOpaqueMask(th, Size, 25, 10, 20, 0x10, false);
                 // calculate the areas of this that need to be overlapped to consider this covered (exclude shadow)
-                this.ContentMask = GeneralUtils.MakeOpaqueMask(th, this.Size, 25, 10, 20, 0xE0, !Globals.UseClassicFiles); 
+                ContentMask = GeneralUtils.MakeOpaqueMask(th, Size, 25, 10, 20, 0xE0, !Globals.UseClassicFiles); 
             }
             else
             {
-                this.Thumbnail = null;
+                Thumbnail = null;
             }
             if (oldImage != null)
             {
@@ -230,8 +232,8 @@ namespace MobiusEditor.Model
 
         public void Reset()
         {
-            Bitmap oldImage = this.Thumbnail;
-            this.Thumbnail = null;
+            Bitmap oldImage = Thumbnail;
+            Thumbnail = null;
             if (oldImage != null)
             {
                 try { oldImage.Dispose(); }